third_party/blink/web_tests/external/wpt/css/css-properties-values-api/animation/custom-property-animation-resolution-comma-list-expected.txt

This is a testharness.js-based test.
[FAIL] Animating a custom property of type <resolution># with iterationComposite
  assert_equals: expected "250dppx, 275dppx" but got "50dppx, 75dppx"
Harness: the test ran to completion.
